linux/efa: Verify that the kernel headers are installed before adding to DKMS
When using DKMS we must verify that the kernel headers are available,
and not only rely on the fact that the kernel is installed.

+++ kernel/linux/efa/rpm/efa-gdr.spec
@@ -29,8 +29,10 @@ Conflicts:	efa
 cd %{install_path}
 dkms add -m %{driver_name} -v %{driver_version}
 for kernel in $(/bin/ls /lib/modules); do
-	dkms build -m %{driver_name} -v %{driver_version} -k $kernel
-	dkms install --force -m %{driver_name} -v %{driver_version} -k $kernel
+	if [ -e /lib/modules/$kernel/build/include ]; then
+		dkms build -m %{driver_name} -v %{driver_version} -k $kernel
+		dkms install --force -m %{driver_name} -v %{driver_version} -k $kernel
+	fi
 done
 
 %preun
